Rename factory to lock_provider_factory
