+++ src/python_sdk/locks/__init__.py
@@ -1,6 +1,3 @@
-from python_sdk.locks._factory import AWSDynamoDBConfig as AWSDynamoDBConfig
-from python_sdk.locks._factory import S3Config as S3Config
-from python_sdk.locks._factory import get_lock_provider as get_lock_provider
 from python_sdk.locks._lock_provider import Lock as Lock
 from python_sdk.locks._lock_provider import LockAcquisitionError as LockAcquisitionError
 from python_sdk.locks._lock_provider import LockInfo as LockInfo
@@ -8,3 +5,6 @@ from python_sdk.locks._lock_provider import LockIsPermanent as LockIsPermanent
 from python_sdk.locks._lock_provider import LockNotOwnedByUs as LockNotOwnedByUs
 from python_sdk.locks._lock_provider import LockProvider as LockProvider
 from python_sdk.locks._lock_provider import LockTaken as LockTaken
+from python_sdk.locks._lock_provider_factory import AWSDynamoDBConfig as AWSDynamoDBConfig
+from python_sdk.locks._lock_provider_factory import S3Config as S3Config
+from python_sdk.locks._lock_provider_factory import get_lock_provider as get_lock_provider
