--- src/python_sdk/locks/_factory/__init__.py
@@ -1,5 +1,0 @@
-from python_sdk.locks._factory._aws_dynamodb_config import AWSDynamoDBConfig as AWSDynamoDBConfig
-from python_sdk.locks._factory._aws_dynamodb_config import AWSDynamoDBConfigDict as AWSDynamoDBConfigDict
-from python_sdk.locks._factory._factory import get_lock_provider as get_lock_provider
-from python_sdk.locks._factory._s3_config import S3Config as S3Config
-from python_sdk.locks._factory._s3_config import S3ConfigDict as S3ConfigDict
