+++ src/python_sdk/locks/_lock_provider_factory/_aws_dynamodb_config.py
